Extend draw_tree_structure so that one can choose whether all the bboxes should be outputed into a single file or every layer to the separate file.

+++ src/appleseed/renderer/kernel/lighting/lighttree_node.h
@@ -65,7 +65,7 @@ class LightTreeNode
         return m_node_luminance;
     }
 
-    float get_level() const
+    size_t get_level() const
     {
         return m_node_level;
     }
